Rename the PlatformPlatform.SharedKernel.Identity namespace to IdGenerators
