+++ application/account-management/Core/Tenants/Domain/TenantTypes.cs
@@ -1,4 +1,4 @@
-using PlatformPlatform.SharedKernel.Identity;
+using PlatformPlatform.SharedKernel.IdGenerators;
 
 namespace PlatformPlatform.AccountManagement.Core.Tenants.Domain;
