+++ application/account-management/Core/Users/Domain/UserTypes.cs
@@ -1,4 +1,4 @@
-using PlatformPlatform.SharedKernel.Identity;
+using PlatformPlatform.SharedKernel.IdGenerators;
 
 namespace PlatformPlatform.AccountManagement.Core.Users.Domain;
