+++ application/shared-kernel/SharedKernel/IdGenerators/IdPrefixAttribute.cs
@@ -1,4 +1,4 @@
-namespace PlatformPlatform.SharedKernel.Identity;
+namespace PlatformPlatform.SharedKernel.IdGenerators;
 
 [AttributeUsage(AttributeTargets.Class)]
 public sealed class IdPrefixAttribute(string prefix) : Attribute
