+++ application/shared-kernel/Tests/Entities/EntityTests.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using PlatformPlatform.SharedKernel.Entities;
-using PlatformPlatform.SharedKernel.Identity;
+using PlatformPlatform.SharedKernel.IdGenerators;
 using Xunit;
 
 namespace PlatformPlatform.SharedKernel.Tests.Entities;
