+++ application/shared-kernel/Tests/Identity/StronglyTypedUlidTests.cs
@@ -1,5 +1,5 @@
 using FluentAssertions;
-using PlatformPlatform.SharedKernel.Identity;
+using PlatformPlatform.SharedKernel.IdGenerators;
 using Xunit;
 
 namespace PlatformPlatform.SharedKernel.Tests.Identity;
